requirements.txt
Flask_Cors==3.0.10
Flask==1.1.2
python-dotenv==0.17.1
gunicorn==20.1.0
